Moving optim kernels to tensor ops
