+++ examples/rl-ppo.rs
@@ -2,11 +2,7 @@
 
 use std::time::Instant;
 
-use dfdx::{
-    optim::{Momentum, Sgd, SgdConfig},
-    prelude::*,
-    tensor::AutoDevice,
-};
+use dfdx::{optim::Sgd, prelude::*, tensor::AutoDevice};
 
 const BATCH: usize = 64;
 const STATE: usize = 4;
