+++ src/optim/mod.rs
@@ -35,11 +35,10 @@ mod optimizer;
 mod rmsprop;
 mod sgd;
 
-pub use adam::{Adam, AdamConfig, AdamKernel};
-pub use optimizer::{Momentum, WeightDecay};
+pub use adam::Adam;
 pub use optimizer::{Optimizer, OptimizerUpdateError, UnusedTensors};
-pub use rmsprop::{RMSprop, RMSpropConfig, RMSpropKernel};
-pub use sgd::{Sgd, SgdConfig, SgdKernel};
+pub use rmsprop::RMSprop;
+pub use sgd::Sgd;
 
 pub mod prelude {
     pub use super::{Optimizer, OptimizerUpdateError, UnusedTensors};
